+++ scripts/UFOconvert
@@ -7,8 +7,7 @@ __license__ = 'Released under the MIT License (http://opensource.org/licenses/MI
 __author__ = 'David Raymond'
 __version__ = '1.0.0'
 
-from silfont.genlib import *
-from silfont.UFOlib import *
+from silfont.ufo.ufo import *
 
 argspec = [
     ('ifont',{'help': 'Input font file'}, {'type': 'infont'}),
@@ -16,7 +15,6 @@ argspec = [
     ('-l','--log',{'help': 'Log file'}, {'type': 'outfile', 'def': '_conv.log'})]
 
 def doit(args) :
-    font = args.ifont
     return args.ifont
 
 execute("PSFU",doit, argspec)
